+++ lib/src/models/doc_save_options_data.dart
@@ -134,6 +134,12 @@ class DocSaveOptionsData extends SaveOptionsData {
       imlRenderingMode = null;
     }
 
+    if (json.containsKey('UpdateAmbiguousTextFont')) {
+      updateAmbiguousTextFont = json['UpdateAmbiguousTextFont'] as bool;
+    } else {
+      updateAmbiguousTextFont = null;
+    }
+
     if (json.containsKey('UpdateCreatedTimeProperty')) {
       updateCreatedTimeProperty = json['UpdateCreatedTimeProperty'] as bool;
     } else {