+++ lib/src/models/html_save_options_data.dart
@@ -396,6 +396,12 @@ class HtmlSaveOptionsData extends SaveOptionsData {
       imlRenderingMode = null;
     }
 
+    if (json.containsKey('UpdateAmbiguousTextFont')) {
+      updateAmbiguousTextFont = json['UpdateAmbiguousTextFont'] as bool;
+    } else {
+      updateAmbiguousTextFont = null;
+    }
+
     if (json.containsKey('UpdateCreatedTimeProperty')) {
       updateCreatedTimeProperty = json['UpdateCreatedTimeProperty'] as bool;
     } else {